# perf: refactor to lazy interface loading