Relates to https://github.com/commaai/opendbc/issues/1184

By moving the `car_helpers.py` imports to a `pytest.fixture`, we can shorten the collection time.


Refactors the `car` tests to use:
* a pytest `fixture` for the interface loading (this import is not executed during collection)
* ~`pytest.mark.parametrize` instead of the `parametrize` in select places to enable properly using the fixture.~ coincidentally already fixed in https://github.com/commaai/opendbc/pull/1930

I tried to keep the diff as small as possible.

There are 3 big chunks of time spent during the pytest collection on a fresh run (after compilation):
* first import of `CANDefine`, aka the packer_pyx compiled library (only slow the first run after compilation)
* import of the interfaces in `car_helpers.py` (always slow)
* import of pytest plugins

"Fresh" denotes a recompilation of the `cython` modules.


## Collect only 

`pytest opendbc/car/tests --collect-only` 

| State  | Fresh           | Second Run         |
|--------|-----------------|-------------------|
| master  | 0.56s  | 0.27s  |
| PR | 0.16s  | 0.13s  |

compare to

`pytest opendbc/can/tests --collect-only` 

| State  | Fresh           | Second Run         |
|--------|-----------------|-------------------|
| master  | 0.32s | 0.02s  |
| PR | 0.29s  | 0.04s  |

## Collection + execution

`pytest opendbc/car/tests`

| State  | Fresh           | Second Run         |
|--------|-----------------|-------------------|
| master  | 6.42s   | 6.19s |
| PR | 5.92s | 5.76s |